# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "EEducation API service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:1234",
    "basePath": "/",
    "paths": {
...
        "/admin/profiles/permissions": {
            "patch": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Set user's permission lvl",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Set user's permission lvl \\\\ ADMIN ONLY",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.SetPermissionLevelReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal"
                    }
                }
            }
        },
...
        "/login": {
            "post": {
                "description": "Login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Invalid credentinals"
                    },
                    "404": {
                        "description": "User with this email not found"
                    },
                    "500": {
                        "description": "Internal"
                    }
                }
            }
        },
        "/profile/change-password": {
            "patch": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Update user password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update user password",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UpdatePasswordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Incorrect old password"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal"
                    }
                }
            }
        },
...
        "/reg": {
            "post": {
                "description": "Registgration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Register new user",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.RegisterReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "409": {
                        "description": "User already exists"
                    },
                    "500": {
                        "description": "Internal"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "user.LoginReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "user.RegisterReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "user.SetPermissionLevelReq": {
            "type": "object",
            "properties": {
                "initiator_id": {
                    "type": "integer"
                },
                "permission_level": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "user.UpdatePasswordReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string"
                },
                "old_password": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "TokenAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}